Issue #107 - Implement Operational Readiness Award
Initial implementation added utilising a modified version of the OV / Ranged type award. This new version allows for a "baseAward" i.e. just the plain ORA in this instance and sets the initial award count to -1 to get around the fact that all other awards are typically set to 0 upon initialisation.
-1 = No Award
0 = Base Award / ORA
1 = ORA-1C
... ect....